Changed all ints to longs (this is a surprise tool that will help us later)

+++ src/admin/Admin.java
@@ -24,12 +24,12 @@ import main.ModulePriorityQueue;
 public class Admin extends Application{
     private Stage globalStage;
 
-    public static main.ModulePriorityQueue<String, Integer> modulePQ = new ModulePriorityQueue<String, Integer>();
+    public static main.ModulePriorityQueue<String, Long> modulePQ = new ModulePriorityQueue<String, Long>();
     public static SortedSet<String> moduleSet = new TreeSet<String>();
     public static Map<String, String> moduleTitles = new HashMap<>();
     public static Map<String, String> moduleDescriptions = new HashMap<>();
-    public static Map<Pair<String, String>, Integer> timeVacancy= new HashMap<>();
-    public static Map<Pair<String, String>, Integer> timeTotal= new HashMap<>();
+    public static Map<Pair<String, String>, Long> timeVacancy= new HashMap<>();
+    public static Map<Pair<String, String>, Long> timeTotal= new HashMap<>();
 
 
     public static SortedSet<String> studentSet = new TreeSet<String>();
@@ -101,14 +101,14 @@ public class Admin extends Application{
         while(scanner.hasNext()){
             String line = scanner.next();
             String[] args = line.split(",");
-            modulePQ.addTimeSlot(args[0], Integer.parseInt(args[1]), args[2]);
+            modulePQ.addTimeSlot(args[0], Long.parseLong(args[1]), args[2]);
             moduleSet.add(args[0]);
             timeVacancy.put(new Pair<String, String>(args[0],
-                    modulePQ.timeSlotDescriptionMap.get(new Pair<>(args[0], Integer.parseInt(args[1])))),
-                    Integer.parseInt(args[3].strip()));
+                    modulePQ.timeSlotDescriptionMap.get(new Pair<>(args[0], Long.parseLong(args[1])))),
+                    Long.parseLong(args[3].strip()));
             timeTotal.put(new Pair<String, String>(args[0],
-                    modulePQ.timeSlotDescriptionMap.get(new Pair<>(args[0], Integer.parseInt(args[1])))),
-                    Integer.parseInt(args[4].strip()));
+                    modulePQ.timeSlotDescriptionMap.get(new Pair<>(args[0], Long.parseLong(args[1])))),
+                    Long.parseLong(args[4].strip()));
         }
         scanner.close();
         scanner = new Scanner(new File("StudentLoginDetails.txt"));
@@ -185,7 +185,7 @@ public class Admin extends Application{
                         // Update choice box
                         selectedModule = (String) newValue;
                         adminAvailableTimeSlotsComboBox.getItems().clear();
-                        for (Integer timeSlotID : modulePQ.getTimeSlotIDs(selectedModule)){
+                        for (Long timeSlotID : modulePQ.getTimeSlotIDs(selectedModule)){
                             adminAvailableTimeSlotsComboBox.getItems().add(modulePQ.getTimeSlot(selectedModule, timeSlotID));
                         }
                         // set module moduleTitle label
@@ -204,7 +204,7 @@ public class Admin extends Application{
                     if (newValue != null) {
 
                         String selectedTimeSlot = (String) newValue;
-                        adminTotalSpotsTextBox.setText(Integer.toString(timeTotal.get(
+                        adminTotalSpotsTextBox.setText(Long.toString(timeTotal.get(
                                 new Pair<>(selectedModule, selectedTimeSlot))));
                     }
                 });
@@ -258,9 +258,9 @@ public class Admin extends Application{
                 String moduleTitle = adminModuleTitleTextBox.getText();
                 String moduleCode = adminModuleCodeTextBox.getText();
                 String selectedTimeSlot = adminAvailableTimeSlotsComboBox.getValue().toString();
-                Integer totalSpots = Integer.valueOf(-1);
+                Long totalSpots = Long.valueOf(-1);
                 try {
-                    totalSpots = Integer.parseInt(adminTotalSpotsTextBox.getText());
+                    totalSpots = Long.parseLong(adminTotalSpotsTextBox.getText());
                 } catch (NumberFormatException e) {
                     e.printStackTrace();
                     System.out.println(e);
@@ -280,8 +280,8 @@ public class Admin extends Application{
                 adminAllModulesListView.getItems().clear();
                 adminAllModulesListView.getItems().addAll(moduleSet);
 
-                ArrayList<Integer> selectedModuleAllTimeSlots =new ArrayList<>(modulePQ.getTimeSlotIDs(selectedModule));
-                for (int timeSlotID : selectedModuleAllTimeSlots){
+                ArrayList<Long> selectedModuleAllTimeSlots = new ArrayList<>(modulePQ.getTimeSlotIDs(selectedModule));
+                for (long timeSlotID : selectedModuleAllTimeSlots){
                     String timeSlotDescription=modulePQ.getTimeSlot(selectedModule,timeSlotID);
                     modulePQ.deleteTimeSlot(selectedModule,timeSlotID);
                     modulePQ.addTimeSlot(moduleCode,timeSlotID,timeSlotDescription);
@@ -362,8 +362,8 @@ public class Admin extends Application{
         FileWriter outputStream1 = new FileWriter("ModulesAndTimeSlots.txt");
         for (String moduleCode : moduleSet) {
             outputStream.write(moduleCode + "," + moduleTitles.get(moduleCode) + "," + moduleDescriptions.get(moduleCode)+"\n");
-            ArrayList<Integer> currentModuleAllTimeSlots = new ArrayList<>(modulePQ.getTimeSlotIDs(moduleCode));
-            for (int timeSlotID : currentModuleAllTimeSlots) {
+            ArrayList<Long> currentModuleAllTimeSlots = new ArrayList<>(modulePQ.getTimeSlotIDs(moduleCode));
+            for (long timeSlotID : currentModuleAllTimeSlots) {
                 String timeSlotDescription = modulePQ.getTimeSlot(moduleCode, timeSlotID);
                 outputStream1.write(moduleCode+","+timeSlotID+","+timeSlotDescription+","
                         +timeVacancy.get(new Pair<>(moduleCode, timeSlotDescription)) + ','
